chore: color mode - extract functions to separate file, add enum
